Fixed rare floating point imprecision when Sequences calculate nested tweens goto

+++ _DOTween.Assembly/DOTween/Sequence.cs
@@ -249,7 +249,8 @@ namespace DG.Tweening
                         }
                     } else {
                         // Nested Tweener/Sequence
-                        float gotoPos = toPos - sequentiable.sequencedPosition;
+//                        float gotoPos = toPos - sequentiable.sequencedPosition;
+                        float gotoPos = (float)((decimal)toPos - (decimal)sequentiable.sequencedPosition);
                         if (gotoPos < 0) gotoPos = 0;
                         Tween t = (Tween)sequentiable;
                         if (!t.startupDone) continue; // since we're going backwards and this tween never started just ignore it
@@ -285,7 +286,8 @@ namespace DG.Tweening
                         }
                     } else {
                         // Nested Tweener/Sequence
-                        float gotoPos = toPos - sequentiable.sequencedPosition;
+//                        float gotoPos = toPos - sequentiable.sequencedPosition;
+                        float gotoPos = (float)((decimal)toPos - (decimal)sequentiable.sequencedPosition);
                         if (gotoPos < 0) gotoPos = 0;
                         Tween t = (Tween)sequentiable;
                         t.isBackwards = false;
